Fix silent bows (and other sounds)
While expanding the visibility API, I forgot about a few side cases. In Minecraft, there are certain sounds that should play for the initiator and the surrounding players. Bows, for example, where not making their sound effects because the source entity was never intended (by Minecraft's standards) to be taken into consideration.